+++ test/Microsoft.AspNetCore.Routing.Tests/TestObjects/TestEndpoint.cs
@@ -1,0 +1,12 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Microsoft.AspNetCore.Routing.Tests.TestObjects
+{
+    internal class TestEndpoint : Endpoint
+    {
+        public TestEndpoint(EndpointMetadataCollection metadata, string displayName) : base(metadata, displayName)
+        {
+        }
+    }
+}
