+++ test/Microsoft.AspNetCore.Routing.Tests/TestObjects/TestServiceProvider.cs
@@ -1,0 +1,15 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+
+namespace Microsoft.AspNetCore.Routing.Tests.TestObjects
+{
+    internal class TestServiceProvider : IServiceProvider
+    {
+        public object GetService(Type serviceType)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
